# openapi.yaml
openapi: 3.0.3
info:
  title: City Pulse API
  version: 0.1.0
  description: >
    REST API for City Pulse — users (city location = San Diego only), regions,
    events, trends (most interacted events by region), and interactions (likes, comments, attending).
    Authenticate via POST /api/auth/login and use the returned access_token as Bearer token for GET /api/auth/me.
servers:
  - url: http://localhost:8000
    description: Local development
paths:
...
  /api/auth/login:
    post:
      summary: Login and get Bearer token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Success; use access_token as Bearer for protected endpoints
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    example: "bearer"
        "401":
          description: Incorrect email or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
  /api/auth/me:
    get:
      summary: Get current authenticated user
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRead"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"

  # ---------------------------------------------------------------------------
  # Users
  # ---------------------------------------------------------------------------
...
  /api/regions/{region_id_or_name}/events:
    get:
      summary: List events in a region
      tags: [Regions]
      parameters:
        - name: region_id_or_name
          in: path
          required: true
          description: Region ID (0) or name ('san diego', 'san-diego')
          schema:
            oneOf:
              - type: integer
                example: 0
              - type: string
                example: "san diego"
      responses:
        "200":
          description: Events in the region
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventRead"
        "400":
          description: Unknown region
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
...
  /api/events:
    get:
      summary: List events (optional region filter)
      tags: [Events]
      parameters:
        - name: region
          in: query
          schema:
            type: string
            default: "san diego"
          description: 'san diego or 0'
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventRead"
        "400":
          description: Unknown region
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
    post:
      summary: Create a new event (in user's region; user must be San Diego)
      tags: [Events]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventCreate"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventRead"
        "400":
          description: User must have city_location san diego / Region not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
  /api/events/{id}:
    get:
      summary: Get event by ID
      tags: [Events]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventRead"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
    put:
      summary: Update event title and/or content
      tags: [Events]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Update success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
    delete:
      summary: Delete event by ID
      tags: [Events]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deletion success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"

  # ---------------------------------------------------------------------------
  # Trends (read-only for users; system maintains list)
  # ---------------------------------------------------------------------------
...
components:
  schemas:
    ErrorDetail:
      type: object
      description: Standard error body
      properties:
        detail:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
                properties:
                  loc:
                    type: array
                    items: {}
                  msg:
                    type: string
                  type:
                    type: string
    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
              msg:
                type: string
              type:
                type: string
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          description: User email (same as when account was created)
        password:
          type: string
          description: Current password
    UserRead:
      type: object
      description: User profile (API response)
      properties:
        id:
          type: integer
          example: 0
        name:
          type: string
          example: "John"
        email:
          type: string
          format: email
          example: "john@example.com"
        city_location:
          type: string
          nullable: true
          example: "san diego"
        created_at:
          type: string
          format: date-time
...
    EventRead:
      type: object
      description: Event in a region
      properties:
        id:
          type: integer
        region_id:
          type: integer
        user_id:
          type: integer
          nullable: true
        title:
          type: string
        content:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    EventCreate:
      type: object
      required: [user_id, title]
      properties:
        user_id:
          type: integer
          description: User posting (must have city_location = san diego)
        title:
          type: string
          minLength: 1
          maxLength: 512
        content:
          type: string
          nullable: true
          maxLength: 10000
    EventUpdate:
      type: object
      properties:
        title:
          type: string
          nullable: true
          minLength: 1
          maxLength: 512
        content:
          type: string
          nullable: true
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Use access_token from POST /api/auth/login